feat: socket support

+++ src/app/fdctl/config.h
@@ -127,7 +127,7 @@ typedef struct {
       ushort listen_port;
       char   xdp_mode[ 8 ];
 
-      uint max_concurrent_connections;
+      ulong max_concurrent_connections;
       uint max_concurrent_connection_ids_per_connection;
       uint max_concurrent_streams_per_connection;
       uint max_concurrent_handshakes;
